COMMERCE-7304 Move Filters to the right package
