+++ modules/apps/commerce/commerce-order-content-web/src/main/java/com/liferay/commerce/order/content/web/internal/frontend/taglib/clay/data/OrderFilterFactoryImpl.java
@@ -12,9 +12,10 @@
  * details.
  */
 
-package com.liferay.commerce.order.content.web.internal.frontend;
+package com.liferay.commerce.order.content.web.internal.frontend.taglib.clay.data;
 
 import com.liferay.commerce.order.content.web.internal.frontend.constants.CommerceOrderDataSetConstants;
+import com.liferay.commerce.order.content.web.internal.frontend.taglib.clay.data.OrderFilterImpl;
 import com.liferay.frontend.taglib.clay.data.Filter;
 import com.liferay.frontend.taglib.clay.data.FilterFactory;
 import com.liferay.portal.kernel.util.ParamUtil;
